Simplify API path
